sprints/sprint-1-llm-foundations/week-2/Project/LLM-tool/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pdfplumber>=0.10.0
